+++ crates/data_components/src/arrow/indexed.rs
@@ -793,6 +793,14 @@ impl TableProvider for IndexedMemTable {
     fn get_column_default(&self, column: &str) -> Option<&Expr> {
         self.inner.get_column_default(column)
     }
+
+    async fn delete_from(
+        &self,
+        state: &dyn Session,
+        filters: Vec<Expr>,
+    ) -> Result<Arc<dyn ExecutionPlan>> {
+        self.inner.delete_from(state, filters).await
+    }
 }
 
 #[async_trait]
@@ -968,9 +976,7 @@ impl crate::delete::DeletionTableProvider for IndexedMemTable {
         state: &dyn Session,
         filters: &[Expr],
     ) -> Result<Arc<dyn ExecutionPlan>> {
-        // Delegate deletion to inner MemTable
-        // Note: Index should be updated after deletion
-        crate::delete::DeletionTableProvider::delete_from(&self.inner, state, filters).await
+        TableProvider::delete_from(self, state, filters.to_vec()).await
     }
 }
