+++ crates/data_components/src/lib.rs
@@ -90,6 +90,7 @@ pub mod imap;
 pub mod index_maintenance;
 pub mod object;
 pub mod poly;
+pub mod update;
 
 #[async_trait]
 pub trait Read: Send + Sync {
